[refactor] for swagger
