+++ api/routes/v1/chunks/controller/getChunk.js
@@ -1,0 +1,17 @@
+const Chunk = require('../../../../../models/v1/chunk/index');
+const asyncErrorWrapper = require('../../../asyncErrorWrapper.js');
+
+const getChunk = asyncErrorWrapper(async (req, res, next) => {
+    const chunk = await Chunk.findOne({_id: req.params.id}); 
+    if (chunk.length == 0) {
+        const err = new Error(`Cannot find ${ req.params.id }`);
+        err.status = 404;
+        throw err;
+    }
+    res.status(200).json({ 
+        message: `Find success [find ${ req.params.id }]`, 
+        data: chunk 
+    });
+});
+
+module.exports = getChunk;
