+++ api/routes/v1/chunks/controller/index.js
@@ -1,0 +1,15 @@
+const createChunk = require("./createChunk");
+const getChunk = require("./getChunk");
+const getChunks = require("./getChunks");
+const updateChunk = require("./updateChunk");
+const deleteChunk = require("./deleteChunk");
+const deleteChunks = require("./deleteChunks");
+
+module.exports = {
+    createChunk,
+    getChunk,
+    getChunks,
+    updateChunk,
+    deleteChunk,
+    deleteChunks
+};
