+++ app.js
@@ -5,16 +5,18 @@ const loggerLoader = require('./loaders/logger');
 const expressLoader = require('./loaders/express');
 const gcStorageLoader = require('./loaders/gcStorage');
 const mongooseLoader = require('./loaders/mongoose');
+const swaggerLoader = require('./loaders/swagger');
 
 const app = express();
 
 gcStorageLoader(app);    
 loggerLoader(app);
+swaggerLoader(app);
 expressLoader(app);    
 
 module.exports = app.listen(app.get('port'), () => {
     if (process.env.NODE_ENV !== 'test')
-        console.log(app.get('port'), '번 포트에서 대기 중..!)');
+        console.log(app.get('port'), '번 포트에서 대기 중..!');
 });
